Se implementan endpoints para compras

+++ src/main/java/com/macv/market/persistence/entity/Cliente.java
@@ -11,7 +11,7 @@ public class Cliente {
 
     private String nombre;
     private String apellidos;
-    private Integer celular;
+    private Long celular;
     private String direccion;
 
     @Column(name = "correo_electronico")
@@ -19,7 +19,5 @@ public class Cliente {
 
     @OneToMany(mappedBy = "cliente")
     private List<Compra> compras;
-
-
 
 }
